refactor: move proxy to proxies/tokens and keep proxy for awhile
